Optionals documentation - fix spelling, more emoji
On ⚡ vs ✨: ✨ is a type and should be used to describe a method that doesn't return anything. Think about it like `void`. ⚡ is a value that an optional can take, think about it like null. ⚡ is used way more than ✨ on this page, but I think this is a good thing - ⚡ was harder for me to remember than ✨.